requirements.txt
python-dateutil==2.8.1
PyGithub==1.55
pandas==1.1.2